score_split(): fix test for being at the beginning of the hunk
Since we are scoring a split *before* line `index`, `index == 0`
indicates a split at the top of the file. (Actually, it only means that
the split is at the top of the hunk. But hopefully the fact that we
include ten lines of diff context means that we will rarely hit the
beginning of a hunk unless the hunk is at the beginning of a file.)